Move controls settings and key binder to separate classes and files
Rename and move the key binder `CMenusKeyBinder` from `menus.h`/`menus_settings.cpp` to `CKeyBinder` in `key_binder.h/cpp`.

Move the `CMenus::DoKeyReader` function to `CKeyBinder` and encapsulate the key binder state.

Move controls settings menu from `menus.h`/`menus_settings.cpp` to `menus_settings_controls.h/cpp`.
