+++ src/game/client/components/menus.h
@@ -17,6 +17,7 @@
 #include <game/client/components/community_icons.h>
 #include <game/client/components/mapimages.h>
 #include <game/client/components/menus_ingame_touch_controls.h>
+#include <game/client/components/menus_settings_controls.h>
 #include <game/client/components/menus_start.h>
 #include <game/client/components/skins7.h>
 #include <game/client/lineinput.h>
@@ -27,20 +28,6 @@
 #include <deque>
 #include <optional>
 #include <vector>
-
-// component to fetch keypresses, override all other input
-class CMenusKeyBinder : public CComponent
-{
-public:
-	const void *m_pKeyReaderId;
-	bool m_TakeKey;
-	bool m_GotKey;
-	IInput::CEvent m_Key;
-	int m_ModifierCombination;
-	CMenusKeyBinder();
-	int Sizeof() const override { return sizeof(*this); }
-	bool OnInput(const IInput::CEvent &Event) override;
-};
 
 class CMenus : public CComponent
 {
@@ -75,14 +62,6 @@ private:
 	void DoLaserPreview(const CUIRect *pRect, ColorHSLA OutlineColor, ColorHSLA InnerColor, int LaserType);
 	int DoButton_GridHeader(const void *pId, const char *pText, int Checked, const CUIRect *pRect, int Align = TEXTALIGN_ML);
 	int DoButton_Favorite(const void *pButtonId, const void *pParentId, bool Checked, const CUIRect *pRect);
-
-	int DoKeyReader(const void *pId, const CUIRect *pRect, int Key, int ModifierCombination, int *pNewModifierCombination);
-
-	void DoSettingsControlsButtons(int Start, int Stop, CUIRect View);
-
-	float RenderSettingsControlsJoystick(CUIRect View);
-	void DoJoystickAxisPicker(CUIRect View);
-	void DoJoystickBar(const CUIRect *pRect, float Current, float Tolerance, bool Active);
 
 	bool m_SkinListScrollToSelected = false;
 	std::optional<std::chrono::nanoseconds> m_SkinList7LastRefreshTime;
@@ -571,14 +550,16 @@ protected:
 	void RenderSettingsTeeCustom7(CUIRect MainView);
 	void RenderSkinSelection7(CUIRect MainView);
 	void RenderSkinPartSelection7(CUIRect MainView);
-	void RenderSettingsControls(CUIRect MainView);
-	void ResetSettingsControls();
 	void RenderSettingsGraphics(CUIRect MainView);
 	void RenderSettingsSound(CUIRect MainView);
 	void RenderSettings(CUIRect MainView);
 	void RenderSettingsCustom(CUIRect MainView);
 
-	std::vector<CButtonContainer> m_vButtonContainersJoystickAbsolute = {{}, {}};
+	// found in menus_settings_controls.cpp
+	// TODO: Change PopupConfirm to avoid using a function pointer to a CMenus
+	//       member function, to move this function to CMenusSettingsControls
+	void ResetSettingsControls();
+
 	std::vector<CButtonContainer> m_vButtonContainersNamePlateShow = {{}, {}, {}, {}};
 	std::vector<CButtonContainer> m_vButtonContainersNamePlateKeyPresses = {{}, {}, {}, {}};
 
@@ -619,8 +600,6 @@ protected:
 
 public:
 	void RenderBackground();
-
-	CMenusKeyBinder m_Binder;
 
 	CMenus();
 	int Sizeof() const override { return sizeof(*this); }
@@ -811,9 +790,11 @@ public:
 
 private:
 	CCommunityIcons m_CommunityIcons;
-	CMenusStart m_MenusStart;
 	CMenusIngameTouchControls m_MenusIngameTouchControls;
 	friend CMenusIngameTouchControls;
+	CMenusSettingsControls m_MenusSettingsControls;
+	friend CMenusSettingsControls;
+	CMenusStart m_MenusStart;
 
 	static int GhostlistFetchCallback(const CFsFileInfo *pInfo, int IsDir, int StorageType, void *pUser);
